feat: 토크해요 페이지에 중복되는 코드가 많아 하나의 컴포넌트로 관리하도록 수정
